Enum dex_pair_io::InnerAction

pub enum InnerAction {
    AddLiquidity {
        amount_a_desired: u128,
        amount_b_desired: u128,
        amount_a_min: u128,
        amount_b_min: u128,
        to: ActorId,
        deadline: u64,
    },
    RemoveLiquidity {
        liquidity: Amount,
        amount_a_min: u128,
        amount_b_min: u128,
        to: ActorId,
        deadline: u64,
    },
    SwapExactTokensForTokens {
        swap_kind: SwapKind,
        amount_in: u128,
        amount_out_min: u128,
        to: ActorId,
        deadline: u64,
    },
    SwapTokensForExactTokens {
        swap_kind: SwapKind,
        amount_out: u128,
        amount_in_max: u128,
        to: ActorId,
        deadline: u64,
    },
    Skim(ActorId),
    Sync,
    Transfer {
        to: ActorId,
        amount: Amount,
    },
}

A part of Action.

Variants

AddLiquidity

Fields

amount_a_desired: u128

An amount of the A tokens to add as liquidity if the B/A price is <= amount_b_desired/amount_a_desired (A depreciates).

amount_b_desired: u128

An amount of the B tokens to add as liquidity if the A/B price is <= amount_a_desired/amount_b_desired (B depreciates).

amount_a_min: u128

Bounds an extent to which the B/A price can go up before this action reverts. Must be <= amount_a_desired.

amount_b_min: u128

Bounds an extent to which the A/B price can go up before this action reverts. Must be <= amount_b_desired.

to: ActorId

A recipient of minted liquidity tokens.

deadline: u64

Timestamp (in ms) after which this action will revert.

Adds liquidity to the contract from [msg::source()]’s fungible tokens and mints liquidity tokens to it.

Requirements

amount_a_desired & amount_b_desired mustn’t equal to 0.

On the first addition (first mint), a resulted amount of pool tokens must be more than MINIMUM_LIQUIDITY.

On success, replies with Event::AddedLiquidity.

RemoveLiquidity

Fields

liquidity: Amount

An amount of liquidity tokens to remove.

amount_a_min: u128

A minimum amount of the A tokens that must be received for this action not to revert.

amount_b_min: u128

A minimum amount of the B tokens that must be received for this action not to revert.

to: ActorId

A recipient of returned fungible tokens.

deadline: u64

Timestamp (in ms) after which this action will revert.

Removes liquidity from the contract by burning [msg::source()]’s liquidity tokens and transferring an appropriate amount of fungible tokens from the contract to it.

Requirements

[msg::source()] must have the same or a greater amount of liquidity tokens than a given one.

On success, replies with Event::RemovedLiquidity.

SwapExactTokensForTokens

Fields

swap_kind: SwapKind

amount_in: u128

amount_out_min: u128

A minimum amount of output tokens that must be received for this action not to revert.

to: ActorId

A recipient of output tokens.

deadline: u64

Timestamp (in ms) after which this action will revert.

Requirements

to mustn’t equal to the contract’s SFT pair.

amount_in mustn’t equal to 0.

On success, replies with Event::Swap.

SwapTokensForExactTokens

Fields

swap_kind: SwapKind

amount_out: u128

amount_in_max: u128

A maximum amount of input tokens that must be received for this action not to revert.

to: ActorId

A recipient of output tokens.

deadline: u64

Timestamp (in ms) after which this action will revert.

Swaps an exact amount of output tokens for as few input tokens as possible.

Requirements

to mustn’t equal to the contract’s SFT pair.

amount_out mustn’t equal to 0.

On success, replies with Event::Swap.

Skim(ActorId)

Tuple Fields

0: ActorId

A recipient of excess tokens.

Syncs the contract’s tokens reserve with actual contract’s balances by transferring excess tokens to some [ActorId].

On success, replies with Event::Skim.

Sync

Syncs the contract’s tokens reserve with actual contract’s balances by setting the reserve equal to the balances.

On success, replies with Event::Sync.

Transfer

Fields

to: ActorId

amount: Amount

Transfers liquidity tokens from [msg::source()].

Requirements

[msg::source()] must have the same or a greater amount of liquidity tokens than a given one.

On success, replies with Event::Transfer.